# perf: use torch.topk to construct nlist