@caic99 2025-05-21
@iProzd Since [`torch.topk`](https://docs.pytorch.org/docs/stable/generated/torch.topk.html) is not guaranteed to be stable, the neighbor atoms with the same distance might have different indices. This fails the UT.